Allow path finding to determine all moves to given destinations
Rather than simply exit on finding a single path to a list of given
destinations, `findPath` has been refactored to use `findPaths` as a delegate.

This means that `findPaths` will return all moves
that will give the shortest path to any destination.

The motivation behind this is allow the snake to make
a decision on which path to take based on how open the region is.