Refactor path-finding to use external flood-fill method (#17)
Refactors the path finding logic into two methods,
`floodFill` which returns a list of all bounded points and
`findPath` which finds a path to a given point.

`findPath` delegates to `floodFill` in order to find a path
to a destination and then determines the move.

This change will allow `floodFill` to also be used to find
enclosed regions.